+++ server/src/main/java/com/abm/pos/ABMPos/manager/VendorManager.java
@@ -2,6 +2,7 @@ package com.abm.pos.ABMPos.manager;
 
 import com.abm.pos.ABMPos.dao.ProductDao;
 import com.abm.pos.ABMPos.dao.VendorDao;
+import com.abm.pos.ABMPos.repository.ProductRepository;
 import com.abm.pos.ABMPos.repository.VendorRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -18,15 +19,40 @@ public class VendorManager {
     @Autowired
     private VendorRepository vendorRepository;
 
+    @Autowired
+    private ProductRepository productRepository;
+
 
     public void addVendor(VendorDao vendorDao) {
 
         vendorRepository.save(vendorDao);
     }
 
-    public void deleteVendor(int vendorId) {
+    public String deleteVendor(int vendorId) {
 
-        vendorRepository.delete(vendorId);
+        boolean isDeletableModel = true;
+        String response = "";
+        try
+        {
+            List<ProductDao> productDaoList = productRepository.findAllByVendorId(String.valueOf(vendorId));
+            if(null != productDaoList && productDaoList.size() > 0)
+            {
+                isDeletableModel = false;
+                System.out.println("Can Not Delete Vendor!!!");
+                response = "Can Not Delete Vendor, Please Delete All Products From This Vendor First";
+            }
+
+            if(isDeletableModel){
+                vendorRepository.delete(vendorId);
+                System.out.println("Vendor Deleted Successfully!!!");
+                response = "Vendor Delete Successfully";
+            }
+        }
+        catch (Exception e){
+            System.out.println("Opps Some Exception is coming!!!");
+        }
+
+        return response;
     }
 
     public List<VendorDao> getVendor() {
